delete lessonListView and move functionality to LessonsContentView
