--- godtools/App/Features/Lessons/Lessons/LessonsListView/LessonsListView.swift
@@ -1,62 +1,0 @@
-//
-//  LessonsListView.swift
-//  godtools
-//
-//  Created by Rachael Skeath on 7/13/22.
-//  Copyright © 2022 Cru. All rights reserved.
-//
-
-import SwiftUI
-
-struct LessonsListView: View {
-    
-    // MARK: - Properties
-    
-    @ObservedObject var viewModel: LessonsListViewModel
-    let width: CGFloat
-    let leadingPadding: CGFloat
-    
-    // MARK: - Body
-    
-    var body: some View {
-        
-        Group {
-            
-            VStack(alignment: .leading, spacing: 5) {
-                
-                Text(viewModel.sectionTitle)
-                    .font(FontLibrary.sfProTextRegular.font(size: 22))
-                    .foregroundColor(ColorPalette.gtGrey.color)
-                
-                Text(viewModel.subtitle)
-                    .font(FontLibrary.sfProTextRegular.font(size: 14))
-                    .foregroundColor(ColorPalette.gtGrey.color)
-            }
-            .padding(.top, 24)
-            .padding(.bottom, 7)
-            .padding([.leading, .trailing], leadingPadding)
-            
-            LessonCardsView(viewModel: viewModel, width: width, leadingPadding: leadingPadding)
-        }
-    }
-}
-
-struct LessonsListView_Previews: PreviewProvider {
-    static var previews: some View {
-        
-        let appDiContainer: AppDiContainer = SwiftUIPreviewDiContainer().getAppDiContainer()
-        
-        let viewModel = LessonsListViewModel(
-            dataDownloader: appDiContainer.initialDataDownloader,
-            localizationServices: appDiContainer.localizationServices,
-            getBannerImageUseCase: appDiContainer.domainLayer.getBannerImageUseCase(),
-            getLanguageAvailabilityUseCase: appDiContainer.domainLayer.getLanguageAvailabilityUseCase(),
-            getLessonsUseCase: appDiContainer.domainLayer.getLessonsUseCase(),
-            getSettingsParallelLanguageUseCase: appDiContainer.domainLayer.getSettingsParallelLanguageUseCase(),
-            getSettingsPrimaryLanguageUseCase: appDiContainer.domainLayer.getSettingsPrimaryLanguageUseCase(),
-            delegate: nil
-        )
-        
-        LessonsListView(viewModel: viewModel, width: 350, leadingPadding: 20)
-    }
-}
